Ahoy! Fisherman 0.3.0
* Fix a critical bug in the Makefile that was
incorrectly merging any existing user configuration
file and the generated Fisherman configuration.
Closes #21.

* Fix a bug in install and uninstall that was adding
plugin names to fishfiles instead of the URL when
interacting with custom URLs. Probably closes #23.

* Fix a bug in install, update and uninstall that
was displaying an incorrect plugin count if there
was at least on failure.

* Fix bug in `fisher install` that causes install
to fail even though it succeeds, due to `wait(1)`'s
behavior of returning `1` if there is any output to
standard error. Closes #20.

* Fix bug in `fisher uninstall` that was removing
plugins from the cache by mistake.

* Add feature to Makefile to download the index for
the first time in order to provide auto-complete
before the user can install/update/search, actions
which would case the index to be updated.

* Add link to Slack [room][wharf] in README. Thanks
@simnalamburt.

* Add new `$fisher_timeout` configuration variable
that lets you specify `curl(1)` `--max-time` option.
Without this, `curl` could hang for a long time if
you are in a bad connection.

* Add `fisher install --link` to allow installing
plugins creating a symbolic link to each of the
relevant files to be copied during the install
process. If you use ***`--link`*** to install a
plugin that is a _path to a directory_ or file, a
symbolic link to the directory will be created making
local testing more convenient as you are not required
to update the plugin's repository to test changes
within Fisherman. If you are testing using
[Fishtape][fishtape] you do not even need to reset
the shell session.

* Add `fisher --alias[=<command>=<alias>]` to simplify
creating new aliases for `fisher` commands. Use
`fisher --alias` without arguments to list the current
set of aliases. Also add auto-complete for aliases
to install, update or uninstall. Note that aliases
are **not** persisted this way. To save your aliases
use `$fisher_alias` as described in `fisher help
config`. Also note that aliases are only auto-complete
if you call `fisher --alias`. To auto-complete aliases
saved to `$fisher_alias` you can do `fisher --alias
(fisher --alias)`.

* Add short options for new and old fisher flags:

    * `--file` → `-f` * `--list` → `-l` * `--alias`
    → `-a`

* Improve help message for failed installs. Closes

* Improve `fisher --validate` to automatically correct
common misspellings, for example when installing a
oh-my-fish package, one often types ohmyifsh.

* :point_up: Improve auto-complete performance by
extracting the implementation of the different
`fisher` flags to `__fisher_*` functions.
`completions/fisher.fish` relies heavily in
`fisher_search` to query what plugins are available
to install/update/uninstall. In this process, numerous
calls to `fisher --list` and `fisher --validate`,
etc., are made. Now, auto-complete does not have to
pay the penalty of entering `fisher`, parsing options,
etc. Closes #27. @namandistro

* Improve `fisher --help` output and show up until
now poorly documented ***`--list`***, ***`--file`***,
etc. flags consistently. Also display available
commands after `make install` to improve usability.

* Improve `fisher install` so that it checks whether
the plugin you are trying to install, if it is already
in the cache, is a symbolic link or not, and installs
it as if the `--link` flag was specified.

* Improve `fisher --validate` to retrieve the absolute
path to the closest directory of the given items if
they are valid local paths. Related #19.

* Improve install to not `git clone` local plugins
if a regular `path/to/file` is given to `fisher
install`. Instead, copy to the cache using `cp(1)`
and if `--link` is used, create a symlink.

* Improve `fisher --validate` to invalidate items
with repeated `.` and `-` and allow items that begin
with `/` or `./` to support installing plugins from
local paths. Related #19.

* Modify `fisher update` default behavior. Now this
command updates Fisherman by default. Use of `--self`
and `--me` is also **deprecated**. To read from the
standard input use a dash `-`. For example: `fisher
--list | fisher update -`. Closes #25.

* Rename `--cache` to more descriptive ***`--list`***.
Thanks @colstrom.

* Remove `fisher --cache=base` and make it return
the base names of all directories in the path by
default. To get the full path use printf `printf
"$fisher_cache/%s" (fisher --list)`

* Rename undocumented `fisher --translate` flag
(again) to `fisher --cache`. This function reads the
standard input for a name, URL or local path and
calculates the plugin's path relative to the cache.
For a name this is simple `$fisher_cache/<name>` for
an URL, retrieve the remote URL of every repository
until there is a match with the given URL and return
the path in the cache of that repository. Finally,
if the input is a

* Revert #3. The reason `getopts.fish` was in its
own file originally is because @bucaran wanted a
standalone, dependency free cli parser solution,
arguably slightly faster than having Awk read
`getopts.awk` for each use. The performance improvement
is negligible at best, but `getopts` is also used
by every single command and future commands and
plugins are very likely to use it as well, so we
so we might as well use the slightly faster version.