Add temporary upgrade check.
New in 0.5.0 is $fisher_file that points to the user fishfile.
Users upgrading from < 0.4.0 will not have this variable set,
which can cause errors whenever Fisherman tries to access it.

This commit adds a simple check to define the variable in the
within the same session and improves the update log message to
display the version Fisherman was updated to.