Add user configuration before sourcing Fisherman configuration
Obviously this should be the other way around, otherwise, the
user would not be able to modify any variables that have an
effect when Fisherman starts.

In other words, when users install Fisherman, the code in
config.fish should be added before sourcing Fisherman
configuration and not after (current way).