Optimize fisher startup path
With the addition of keybinding support for fish 2.x, fisher greatly
increases shell startup time (50-100ms in my testing). This is due to a
number of reasons, including using less-than-ideal builtins,
unconditionally running mkdir, unconditionally recreating
conf.d/fisher.fish, and checking for curl on startup (instead of when
curl is used).

This patch cuts fisher-related startup time to roughly 25ms, simplifying
the code paths as much as I am able.

The source function is removed as a compromise for speed. Fish 2.1 and
greater support `source`, and supporting fish 2.0 is not worth a hefty
speed penalty in my opinion.

curl checks have been moved to the two uses of curl -- there is some
duplication now so this may be slightly less ideal. Also, curl checks
are now done with the builtin `type` instead of `which`, which is not
guaranteed to exist.