self-update: better network error handling
By using curl's -f option, we improve the self-update mechanism by not
overwriting the fisher script if the HTTP status code given to us by
Github is higher than 400.

Additionally, we make sure that the exit code of the curl process must
be always zero for the self-update to succeed.