added support for sourcehut plugin source (#801)
This commit adds support for sr.ht and git.sr.ht
plugin sources. A url to a git.sr.ht repo OR to an
sr.ht project is valid. Additionally, if a plugin
name begins with ~ and is NOT followed by a /,
such as ~username/repo, it will be treated as a
sourcehut source.

Examples of valid sourcehut names:
    ~username/repo
    sr.ht/~username/repo
    git.sr.ht/~username/repo
    https://git.sr.ht/~username/repo

However, the following is still a $HOME expansion:
    ~/username/repo
because the ~ is followed by a /.